requirements.txt
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
httplib2==0.14.0
idna==2.8
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
mysql-connector-python==8.0.27
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.19.1
pycairo==1.16.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
python-apt===2.0.0-ubuntu0.20.04.6-elementary6-ubuntu6.1
python-debian===0.1.36ubuntu1
python-dotenv==0.17.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.3.1
supervisor==4.1.0
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
xkit==0.0.0
